feat: Change from Fn to FnOnce for the tread closure
Because the closure will never be called more than once. And having it
as Fn means that we need to be careful when we move variables into
the closure because the closure could be called multiple times (or at
least that is what Rust thinks). This leads to requirements to double
clone items, or really clone items for every invocation.