feat: Adding the code for the swerve model
Code for a kinematic tree that describes how the different
parts of the robot are connected. And code for the swerve robot
model which incorporates a kinematic tree and the ability
to create a model by adding parts as well as the ability
to calculate the relative positions between parts.